Divided MomentController into MomentCreateController, MomentDeleteController, MomentFilterController and MomentVViewController. Added test files for new controllers
