+++ src/test/java/dev/equipo3/insideout_java/controller/MomentFilterControllerTest.java
@@ -1,0 +1,5 @@
+package dev.equipo3.insideout_java.controller;
+
+public class MomentFilterControllerTest {
+
+}
